backend/analytics/requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.7
clickhouse-connect==0.15.1
cytoolz==1.1.0
eth-hash==0.8.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
hypersync==1.0.0
idna==3.11
lz4==4.4.5
numpy==2.4.4
pandas==3.0.2
parsimonious==0.10.0
psutil==7.2.2
pyarrow==23.0.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
six==1.17.0
StrEnum==0.4.15
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zstandard==0.25.0

fastapi==0.136.0
starlette==0.49.1
strawberry-graphql[fastapi]==0.312.3
uvicorn==0.38.0

